+++ packages/css/src/components/selectable/selectable.template.ts
@@ -20,6 +20,11 @@ export function selectableTemplate({
   onChange,
   info
 }: Selectable<TemplateResult>) {
+  const ariaDescribedBy = [
+    describedById,
+    info?.fixed ? info.id : undefined
+  ].filter(id => !!id).join(' ') || undefined;
+
   return html`
     <div class="dso-selectable">
       <input
@@ -28,7 +33,7 @@ export function selectableTemplate({
         value=${value}
         name=${ifDefined(name)}
         aria-invalid=${ifDefined(invalid)}
-        aria-describedby=${ifDefined(info ? describedById : undefined)}
+        aria-describedby=${ifDefined(ariaDescribedBy)}
         @change=${(e: Event) => onChange?.(e)}
         ?disabled=${disabled}
         ?required=${required}