+++ packages/sources/src/components/form/form.models.ts
@@ -1,5 +1,6 @@
 export * from './models/form-group-checkboxes.model';
 export * from './models/form-group-confirm.model';
+export * from './models/form-group-files.model';
 export * from './models/form-group-input.model';
 export * from './models/form-group-input-number.model';
 export * from './models/form-group-radios.model';